Pass all parameters as template arguments
