+++ src/pollard_rho_one_find_factor.h
@@ -4,11 +4,50 @@
 #include <optional>
 
 #include "common.h"
+#include "logging.h"
+#include "utility.h"
 
 namespace Project {
 
-std::optional<bigint> find_factor_pollard_rho_one(
-    bigint n, size_t max_iterations = 100000,
-    std::vector<bigint> bases = {2, 3, 5, 7, 11});
+template <typename BasesType, size_t bases_count>
+struct PollardRhoOneParams {
+  const size_t max_iterations;
+  const std::array<BasesType, bases_count> bases;
+};
+
+template <PollardRhoOneParams params>
+std::optional<bigint> find_factor_pollard_rho_one(bigint n) {
+  INFO() << "finding factors of " << n << " using pollard rho one";
+
+  if (n % 2 == 0) {
+    return 2;
+  }
+
+  for (bigint base : params.bases) {
+    DEBUG() << "trying base:  " << base;
+    auto current_pow = base;
+    for (bigint i = 1; i < params.max_iterations; i++) {
+      current_pow = powmod(current_pow, i, n);
+      auto current_gcd = gcd(current_pow - 1, n);
+      DEBUG() << "iteration: " << i << " current_pow: " << current_pow
+              << " current_gcd: " << current_gcd;
+      if (current_pow == 1) {
+        break;
+      }
+      if (current_gcd == n) {
+        break;
+      }
+      if (current_gcd > 1) {
+        return current_gcd;
+      }
+    }
+  }
+  WARNING()
+      << "pollard rho one failed for all start values, probably you need to "
+         "use different "
+         "start value, n: "
+      << n;
+  return std::nullopt;
+}
 
 }  // namespace Project
