+++ src/trial_find_factor.h
@@ -7,6 +7,13 @@
 
 namespace Project {
 
-std::optional<bigint> find_factor_trial(bigint n);
+std::optional<bigint> find_factor_trial(bigint n) {
+  for (bigint i = 2; i * i <= n; i++) {
+    if (n % i == 0) {
+      return i;
+    }
+  }
+  return std::nullopt;
+}
 
 }  // namespace Project
