+++ src/trial_primality_test.cpp
@@ -2,6 +2,8 @@
 
 #include <boost/multiprecision/cpp_int.hpp>
 
+#include "common.h"
+
 namespace Project {
 
 bool is_prime_trial(bigint x) {
